back/requirements.txt
flask
flask_sqlalchemy
pymysql
dotenv
openai
dateparser
requests